+++ autokeras/__init__.py
@@ -1,12 +1,10 @@
 from autokeras.auto_model import AutoModel
 from autokeras.hypermodel.base import Block
 from autokeras.hypermodel.base import Head
-from autokeras.hypermodel.base import HyperBlock
 from autokeras.hypermodel.base import Node
-from autokeras.hypermodel.base import Preprocessor
 from autokeras.hypermodel.block import ConvBlock
 from autokeras.hypermodel.block import DenseBlock
-from autokeras.hypermodel.block import EmbeddingBlock
+from autokeras.hypermodel.block import Embedding
 from autokeras.hypermodel.block import Merge
 from autokeras.hypermodel.block import ResNetBlock
 from autokeras.hypermodel.block import RNNBlock
@@ -15,19 +13,19 @@ from autokeras.hypermodel.block import TemporalReduction
 from autokeras.hypermodel.block import XceptionBlock
 from autokeras.hypermodel.head import ClassificationHead
 from autokeras.hypermodel.head import RegressionHead
-from autokeras.hypermodel.hyperblock import ImageBlock
-from autokeras.hypermodel.hyperblock import StructuredDataBlock
-from autokeras.hypermodel.hyperblock import TextBlock
+from autokeras.hypermodel.block import ImageBlock
+from autokeras.hypermodel.block import StructuredDataBlock
+from autokeras.hypermodel.block import TextBlock
 from autokeras.hypermodel.node import ImageInput
 from autokeras.hypermodel.node import Input
 from autokeras.hypermodel.node import StructuredDataInput
 from autokeras.hypermodel.node import TextInput
-from autokeras.hypermodel.preprocessor import FeatureEngineering
-from autokeras.hypermodel.preprocessor import ImageAugmentation
-from autokeras.hypermodel.preprocessor import LightGBM
-from autokeras.hypermodel.preprocessor import Normalization
-from autokeras.hypermodel.preprocessor import TextToIntSequence
-from autokeras.hypermodel.preprocessor import TextToNgramVector
+# from autokeras.hypermodel.preprocessor import FeatureEngineering
+# from autokeras.hypermodel.preprocessor import ImageAugmentation
+# from autokeras.hypermodel.preprocessor import LightGBM
+from autokeras.hypermodel.block import Normalization
+from autokeras.hypermodel.block import TextToIntSequence
+from autokeras.hypermodel.block import TextToNgramVector
 from autokeras.task import ImageClassifier
 from autokeras.task import ImageRegressor
 from autokeras.task import StructuredDataClassifier
